Re-organize code and minor refactors
* new pkg/rest/{turn,toggle}.go files to hold related functions
* remove pull_request from "on" trigger in release.yml workflow
