+++ .github/workflows/release.yml
@@ -2,7 +2,6 @@
 name: release
 
 on:
-  pull_request:
   push:
     # run only against tags
     tags:
